// app/src/generated/api/models/Trip.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: mateja176@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Trip
 */
export interface Trip {
    /**
     * 
     * @type {string}
     * @memberof Trip
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Trip
     */
    destination: string;
    /**
     * 
     * @type {Date}
     * @memberof Trip
     */
    startDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof Trip
     */
    endDate: Date;
    /**
     * 
     * @type {string}
     * @memberof Trip
     */
    comment?: string;
    /**
     * 
     * @type {object}
     * @memberof Trip
     */
    user?: object;
    /**
     * 
     * @type {Date}
     * @memberof Trip
     */
    publishedAt?: Date;
}

export function TripFromJSON(json: any): Trip {
    return TripFromJSONTyped(json, false);
}

export function TripFromJSONTyped(json: any, ignoreDiscriminator: boolean): Trip {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'destination': json['destination'],
        'startDate': (new Date(json['startDate'])),
        'endDate': (new Date(json['endDate'])),
        'comment': !exists(json, 'comment') ? undefined : json['comment'],
        'user': !exists(json, 'user') ? undefined : json['user'],
        'publishedAt': !exists(json, 'published_at') ? undefined : (new Date(json['published_at'])),
    };
}

export function TripToJSON(value?: Trip | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'destination': value.destination,
        'startDate': (value.startDate.toISOString()),
        'endDate': (value.endDate.toISOString()),
        'comment': value.comment,
        'user': value.user,
        'published_at': value.publishedAt === undefined ? undefined : (value.publishedAt.toISOString()),
    };
}